+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/InstrumentImpl.java
@@ -12,11 +12,11 @@
  *
  * MISO is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with MISO.  If not, see <http://www.gnu.org/licenses/>.
+ * along with MISO. If not, see <http://www.gnu.org/licenses/>.
  *
  * *********************************************************************
  */
@@ -24,19 +24,71 @@
 package uk.ac.bbsrc.tgac.miso.core.data.impl;
 
 import java.io.IOException;
-import java.io.Serializable;
+import java.util.Collections;
+import java.util.Date;
+import java.util.HashSet;
+import java.util.Objects;
+import java.util.Set;
+import java.util.stream.Collectors;
 
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.OneToOne;
 import javax.persistence.Table;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
+import javax.persistence.Transient;
 
-import uk.ac.bbsrc.tgac.miso.core.data.AbstractInstrument;
+import uk.ac.bbsrc.tgac.miso.core.data.Instrument;
 import uk.ac.bbsrc.tgac.miso.core.data.InstrumentModel;
+import uk.ac.bbsrc.tgac.miso.core.data.InstrumentPosition;
+import uk.ac.bbsrc.tgac.miso.core.data.Run;
+import uk.ac.bbsrc.tgac.miso.core.data.ServiceRecord;
+import uk.ac.bbsrc.tgac.miso.core.util.LimsUtils;
 
 @Entity
 @Table(name = "Instrument")
-public class InstrumentImpl extends AbstractInstrument implements Serializable {
+public class InstrumentImpl implements Instrument {
 
   private static final long serialVersionUID = 1L;
+
+  private static final Long UNSAVED_ID = 0L;
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.AUTO)
+  @Column(name = "instrumentId")
+  private long id = InstrumentImpl.UNSAVED_ID;
+
+  @Column(nullable = false)
+  private String name;
+
+  @ManyToOne(targetEntity = InstrumentModel.class)
+  @JoinColumn(name = "instrumentModelId", nullable = false)
+  private InstrumentModel instrumentModel;
+
+  private String serialNumber;
+  @Temporal(TemporalType.DATE)
+  private Date dateCommissioned;
+  @Temporal(TemporalType.DATE)
+  private Date dateDecommissioned = null;
+
+  @OneToOne(targetEntity = InstrumentImpl.class, optional = true)
+  @JoinColumn(name = "upgradedInstrumentId")
+  private Instrument upgradedInstrument;
+
+  @ManyToOne
+  @JoinColumn(name = "defaultPurposeId")
+  private RunPurpose defaultRunPurpose;
+
+  @Transient
+  private Date lastServicedDate;
 
   public InstrumentImpl(String name, InstrumentModel instrumentModel) {
     setName(name);
@@ -51,6 +103,220 @@ public class InstrumentImpl extends AbstractInstrument implements Serializable {
   public InstrumentImpl() {
     setInstrumentModel(null);
     setName(null);
+  }
+
+  @Override
+  public void setId(long id) {
+    this.id = id;
+  }
+
+  @Override
+  public long getId() {
+    return this.id;
+  }
+
+  @Override
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  @Override
+  public String getName() {
+    return this.name;
+  }
+
+  @Override
+  public void setInstrumentModel(InstrumentModel instrumentModel) {
+    this.instrumentModel = instrumentModel;
+  }
+
+  @Override
+  public InstrumentModel getInstrumentModel() {
+    return this.instrumentModel;
+  }
+
+  @Override
+  public void setSerialNumber(String serialNumber) {
+    this.serialNumber = serialNumber;
+  }
+
+  @Override
+  public String getSerialNumber() {
+    return serialNumber;
+  }
+
+  @Override
+  public void setDateCommissioned(Date date) {
+    this.dateCommissioned = date;
+  }
+
+  @Override
+  public Date getDateCommissioned() {
+    return dateCommissioned;
+  }
+
+  @Override
+  public void setDateDecommissioned(Date date) {
+    this.dateDecommissioned = date;
+  }
+
+  @Override
+  public Date getDateDecommissioned() {
+    return dateDecommissioned;
+  }
+
+  @Override
+  public void setUpgradedInstrument(Instrument instrument) {
+    this.upgradedInstrument = instrument;
+  }
+
+  @Override
+  public Instrument getUpgradedInstrument() {
+    return upgradedInstrument;
+  }
+
+  @Override
+  public RunPurpose getDefaultRunPurpose() {
+    return defaultRunPurpose;
+  }
+
+  @Override
+  public void setDefaultRunPurpose(RunPurpose defaultRunPurpose) {
+    this.defaultRunPurpose = defaultRunPurpose;
+  }
+
+  @Override
+  public String toString() {
+    return "AbstractInstrument [id=" + id
+        + ", name=" + name
+        + ", instrumentModel=" + instrumentModel.getId()
+        + ", serialNumber=" + serialNumber
+        + ", dateCommissioned=" + dateCommissioned
+        + ", dateDecommissioned=" + dateDecommissioned
+        + ", upgradedInstrument=" + (upgradedInstrument == null ? null : upgradedInstrument.getId()) + "]";
+  }
+
+  @Override
+  public boolean isActive() {
+    return dateDecommissioned == null;
+  }
+
+  @Override
+  public void setLastServicedDate(Date date) {
+    this.lastServicedDate = date;
+  }
+
+  @Override
+  public Date getLastServicedDate() {
+    return lastServicedDate;
+  }
+
+  @OneToMany(targetEntity = Run.class, mappedBy = "sequencer")
+  private Set<Run> runs = new HashSet<>();
+
+  @Override
+  public Set<Run> getRuns() {
+    return runs;
+  }
+
+  @Override
+  public void setRuns(Set<Run> runs) {
+    this.runs = runs;
+  }
+
+  @OneToMany(targetEntity = ServiceRecord.class, mappedBy = "instrument", cascade = CascadeType.REMOVE)
+  private Set<ServiceRecord> serviceRecords = new HashSet<>();
+
+  @Override
+  public Set<ServiceRecord> getServiceRecords() {
+    return serviceRecords;
+  }
+
+  @Override
+  public void setServiceRecords(Set<ServiceRecord> serviceRecords) {
+    this.serviceRecords = serviceRecords;
+  }
+
+  @Override
+  public boolean isOutOfService() {
+    if (getServiceRecords() == null) {
+      return false;
+    }
+    return getServiceRecords().stream().anyMatch(sr -> sr.isOutOfService() && sr.getEndTime() == null && sr.getStartTime() != null
+        && sr.getStartTime().before(new Date()) && sr.getPosition() == null);
+  }
+
+  @Override
+  public Set<InstrumentPosition> getOutOfServicePositions() {
+    if (getInstrumentModel() == null) {
+      return Collections.emptySet();
+    }
+    if (isOutOfService()) {
+      return getInstrumentModel().getPositions();
+    }
+    return getInstrumentModel().getPositions().stream()
+        .filter(pos -> getServiceRecords().stream().anyMatch(sr -> sr.isOutOfService() && sr.getEndTime() == null
+            && sr.getStartTime() != null && sr.getStartTime().before(new Date()) && sr.getPosition().getAlias().equals(pos.getAlias())))
+        .collect(Collectors.toSet());
+  }
+
+  @Override
+  public String getOutOfServicePositionsLabel() {
+    if (isOutOfService()) {
+      return "All positions";
+    }
+    Set<InstrumentPosition> positions = getOutOfServicePositions();
+    if (positions.isEmpty()) {
+      return null;
+    } else {
+      return "Position" + (positions.size() == 1 ? "" : "a") + " "
+          + positions.stream().map(InstrumentPosition::getAlias).collect(Collectors.joining(", "));
+    }
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(getDateCommissioned(),
+        getDateDecommissioned(),
+        getId(),
+        getLastServicedDate(),
+        getName(),
+        getInstrumentModel(),
+        getRuns(),
+        getSerialNumber(),
+        getServiceRecords(),
+        getUpgradedInstrument(),
+        getDefaultRunPurpose());
+  }
+
+  @Override
+  public boolean equals(Object obj) {
+    return LimsUtils.equals(this, obj, InstrumentImpl::getDateCommissioned,
+        InstrumentImpl::getDateDecommissioned,
+        InstrumentImpl::getId,
+        InstrumentImpl::getLastServicedDate,
+        InstrumentImpl::getName,
+        InstrumentImpl::getInstrumentModel,
+        InstrumentImpl::getRuns,
+        InstrumentImpl::getSerialNumber,
+        InstrumentImpl::getServiceRecords,
+        InstrumentImpl::getUpgradedInstrument,
+        InstrumentImpl::getDefaultRunPurpose);
+  }
+
+  @Override
+  public boolean isSaved() {
+    return getId() != UNSAVED_ID;
+  }
+
+  @Override
+  public String getDeleteType() {
+    return "Instrument";
+  }
+
+  @Override
+  public String getDeleteDescription() {
+    return getName() + " (" + getInstrumentModel().getAlias() + ")";
   }
 
 }