+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/RunPurposeService.java
@@ -1,0 +1,7 @@
+package uk.ac.bbsrc.tgac.miso.core.service;
+
+import uk.ac.bbsrc.tgac.miso.core.data.impl.RunPurpose;
+
+public interface RunPurposeService extends DeleterService<RunPurpose>, ListService<RunPurpose>, SaveService<RunPurpose> {
+
+}